Update Makefiles to use clang-16. divsufsort:         Apply the Debian patches to divsufsort.         Compile and install 32- and 64-bit versions of the library for ec2 instances. Update SA algorithm use to use 32/64 bit distinct types, depending on the number of samples in the data set.         This tool now supports larger data files.                 The previous limit was INT_MAX-1 symbols.                 The new limit is INT64_MAX-1 symbols.                 In general:                         The 32-bit approach requires 13n bytes of memory.                         The 64-bit approach requires 25n bytes of memory. fancymath:         Added uint128_t types (making use of the internal unsigned __int128 type).                 Include support functions:                         Added safe addition routine for 128-bit integers.                         A uint128_t -> string tool for reporting. Tools:         interleave-data:                 Interleave data and output it.

+++ src/entlib.h
@@ -106,8 +106,8 @@ double compressionEstimate(const statData_t *S, size_t L, struct compResult *res
 struct SAresult {
   bool done;
   bool tTupleDone;
-  int32_t u;  // t = u-1
-  int32_t v;
+  size_t u;  // t = u-1
+  size_t v;
   double tTuplePmax;  // Take the upper 99% confidence interval of this parameter
   double tTuplePu;
   double tTupleEntropy;
